fixed the break/deliver/return problem
fixed the problem with break/deliver in if clauses.

No, you can define "syntactic functions" which do
not open a FUNCTION_FRAME. The point is, such FUNCTION_FRAMEs
will block break and deliver.

Also removed the ordering that you cannot 'deliver' out of
a 'loop' (or 'throw'), which seemed unnatural